clean(api): move LearningContentResourceNotFound error in shared/domain/errors.js
