+++ api/src/devcomp/infrastructure/datasources/learning-content/module-datasource.js
@@ -2,7 +2,7 @@ import { readdir } from 'node:fs/promises';
 import { dirname, join } from 'node:path';
 import { fileURLToPath, pathToFileURL } from 'node:url';
 
-import { LearningContentResourceNotFound } from '../../../../shared/infrastructure/datasources/learning-content/LearningContentResourceNotFound.js';
+import { LearningContentResourceNotFound } from '../../../../shared/domain/errors.js';
 
 const referential = await importModules();
