--- api/src/shared/infrastructure/datasources/learning-content/LearningContentResourceNotFound.js
@@ -1,8 +1,0 @@
-class LearningContentResourceNotFound extends Error {
-  constructor({ skillId } = {}) {
-    super();
-    this.skillId = skillId;
-  }
-}
-
-export { LearningContentResourceNotFound };
